Refactor cyral_sidecar
