+++ cyral/internal/integration/teams/resource.go
@@ -2,11 +2,9 @@ package teams
 
 import (
 	"fmt"
-	"net/http"
 
 	"github.com/cyralinc/terraform-provider-cyral/cyral/client"
 	"github.com/cyralinc/terraform-provider-cyral/cyral/core"
-	"github.com/cyralinc/terraform-provider-cyral/cyral/core/types/operationtype"
 	"github.com/cyralinc/terraform-provider-cyral/cyral/core/types/resourcetype"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )
@@ -19,17 +17,6 @@ var resourceContextHandler = core.DefaultContextHandler{
 	BaseURLFactory: func(d *schema.ResourceData, c *client.Client) string {
 		return fmt.Sprintf("https://%s/v1/integrations/notifications/teams", c.ControlPlane)
 	},
-}
-
-var ReadMsTeamsConfig = core.ResourceOperationConfig{
-	ResourceName: resourceName,
-	Type:         operationtype.Read,
-	HttpMethod:   http.MethodGet,
-	URLFactory: func(d *schema.ResourceData, c *client.Client) string {
-		return fmt.Sprintf("https://%s/v1/integrations/notifications/teams/%s", c.ControlPlane, d.Id())
-	},
-	SchemaWriterFactory: func(_ *schema.ResourceData) core.SchemaWriter { return &MsTeamsIntegration{} },
-	RequestErrorHandler: &core.ReadIgnoreHttpNotFound{ResName: "Integration Teams"},
 }
 
 func resourceSchema() *schema.Resource {
