+++ cyral/internal/sidecar/constants.go
@@ -1,0 +1,7 @@
+package sidecar
+
+const (
+	resourceName             = "cyral_sidecar"
+	dataSourceIdName         = "cyral_sidecar_id"
+	dataSourceBoundPortsName = "cyral_sidecar_bound_ports"
+)
