+++ cyral/internal/sidecar/schema_loader.go
@@ -1,0 +1,31 @@
+package sidecar
+
+import (
+	"github.com/cyralinc/terraform-provider-cyral/cyral/core"
+)
+
+type packageSchema struct {
+}
+
+func (p *packageSchema) Name() string {
+	return "sidecar"
+}
+
+func (p *packageSchema) Schemas() []*core.SchemaDescriptor {
+	return []*core.SchemaDescriptor{
+		// {
+		// 	Name:   dataSourceName,
+		// 	Type:   core.DataSourceSchemaType,
+		// 	Schema: dataSourceSchema,
+		// },
+		{
+			Name:   resourceName,
+			Type:   core.ResourceSchemaType,
+			Schema: resourceSchema,
+		},
+	}
+}
+
+func PackageSchema() core.PackageSchema {
+	return &packageSchema{}
+}
